[testsuite] Split Obj-C foundation test
TestObjCMethods2.py was the third-longest running test on Darwin. By
splitting it up, lit can exploit parallelism to reduce the total wall
clock time.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@358088 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit bd94e351bd7a7c2ce6c83e85a5328fc292bae226)
